Remove asyncore, asynchat, and smptd
